make genesis link statically
Though we can use gensis_LDFLAGS = -all-static to get libtool to link
genesis statically, it is insufficient as g++ will implictly try to link
against libstdc++, which will fail due to an attempt to create a dynamic
link in a static binary.

Use -print-file from g++ to get the path to libstdc++.a and then link
against that explicitly using gcc.  This means we need to hand roll the
link command for genesis.

Use pattern substitution to get the .a file for libpcre++ whilst it
remains a dependency.