Basic logging support
We don't have any log destination support atm so for now all log
messages goes to console.

We need support for logging to console, file and syslog at the very
least. We most likely need to support multiple simultaneous destinations
with different loglevels so INFO and higher messages can go to a file
and ERR or higher can go to syslog for example.

We can implement support for this now but it won't be useful until we
have proper config support so might as well wait for that.